expl_hexagon: Subcrate for hexagon mesh
Move asset setup from map to terrain module
