+++ crates/expl_hexagon/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "expl_hexagon"
+version = "0.1.0"
+edition = "2021"
+license = "ISC"
+
+[dependencies]
+bevy_render = "0.11"
+glam = "0.24"
+hexasphere = "9"
