--- src/map/asset.rs
@@ -1,6 +1,0 @@
-use bevy::prelude::*;
-
-#[derive(Resource)]
-pub struct HexAssets {
-    pub mesh: Handle<Mesh>,
-}
