+++ src/map/mod.rs
@@ -1,12 +1,9 @@
-mod asset;
 mod command;
 mod component;
 mod event;
-mod hex;
 mod plugin;
 mod system;
 
-pub use asset::*;
 pub use command::*;
 pub use component::*;
 pub use event::*;
