+++ src/map/system.rs
@@ -1,14 +1,5 @@
-use super::{asset::*, command::*, component::*, event::*, hex::*};
+use super::{command::*, component::*, event::*};
 use bevy::prelude::*;
-
-pub fn insert_hex_assets(mut commands: Commands, mut meshes: ResMut<Assets<Mesh>>) {
-    commands.insert_resource(HexAssets {
-        mesh: meshes.add(Mesh::from(Hexagon {
-            radius: 1.0,
-            subdivisions: 2,
-        })),
-    });
-}
 
 pub fn log_moves(
     mut map_events: EventReader<MapEvent>,
